+++ src/main/java/com/example/demo/entity/PokemonCardDeck.java
@@ -1,4 +1,4 @@
-package com.example.demo.pokemon;
+package com.example.demo.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
